Fix sloppy argument parsing
